🎨 Moved login to its own component, removed old membership
