+++ server/.strapi-updater.json
@@ -1,4 +1,4 @@
 {
 	"latest": "3.3.4",
-	"lastUpdateCheck": 1607639467162
+	"lastUpdateCheck": 1607829093909
 }
