+++ server/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "12/11/2020 4:02:08 AM"
+    "x-generation-date": "12/13/2020 3:11:37 AM"
   },
   "x-strapi-config": {
     "path": "/documentation",
